Migrated League Standing and Scoring
- Migrated League Standing and Scoring generateStaticParams
- Converted generateStaticParams into a single function
